Move get_email_base_context to core
