+++ saleor/dashboard/emails.py
@@ -9,6 +9,7 @@ from django.utils.http import urlsafe_base64_encode
 from templated_email import send_templated_mail
 
 from ..core.utils import build_absolute_uri
+from ..core.utils.email import get_email_base_context
 
 
 @shared_task
@@ -20,7 +21,7 @@ def send_set_password_email(staff):
             'account:reset-password-confirm',
             kwargs={'token': token, 'uidb64': uid}))
     ctx = get_email_base_context()
-    ctx.update({'password_set_url': password_set_url})
+    ctx['password_set_url'] = password_set_url
     send_templated_mail(
         template_name='dashboard/staff/set_password',
         from_email=settings.DEFAULT_FROM_EMAIL,
@@ -31,19 +32,9 @@ def send_set_password_email(staff):
 @shared_task
 def send_promote_customer_to_staff_email(staff):
     ctx = get_email_base_context()
-    ctx.update({'url': build_absolute_uri(reverse('dashboard:index'))})
+    ctx['url'] = build_absolute_uri(reverse('dashboard:index'))
     send_templated_mail(
         template_name='dashboard/staff/promote_customer',
         from_email=settings.DEFAULT_FROM_EMAIL,
         recipient_list=[staff.email],
         context=ctx)
-
-
-def get_email_base_context():
-    site = Site.objects.get_current()
-    logo_url = build_absolute_uri(
-        location=None) + static('images/logo-document.svg')
-    return {
-        'domain': site.domain,
-        'logo_url': logo_url,
-        'site_name': site.name}
