+++ saleor/order/emails.py
@@ -6,6 +6,7 @@ from django.urls import reverse
 from templated_email import send_templated_mail
 
 from ..core.utils import build_absolute_uri
+from ..core.utils.email import get_email_base_context
 from ..seo.schema.email import get_order_confirmation_markup
 from .models import Fulfillment, Order
 
@@ -14,17 +15,6 @@ CONFIRM_FULFILLMENT_TEMPLATE = 'order/confirm_fulfillment'
 UPDATE_FULFILLMENT_TEMPLATE = 'order/update_fulfillment'
 CONFIRM_PAYMENT_TEMPLATE = 'order/payment/confirm_payment'
 CONFIRM_NOTE_TEMPLATE = 'order/note/confirm_note'
-
-
-def get_email_context(order_token):
-    """Prepares context required for email template rendering."""
-    site = Site.objects.get_current()
-    logo_url = build_absolute_uri(
-        location=None) + static('images/logo-document.svg')
-    order_url = build_absolute_uri(
-        reverse('order:details', kwargs={'token': order_token}))
-    ctx = {'site_name': site.name, 'logo_url': logo_url, 'url': order_url}
-    return ctx
 
 
 def collect_data_for_email(order_pk, template):
@@ -36,7 +26,9 @@ def collect_data_for_email(order_pk, template):
     """
     order = Order.objects.get(pk=order_pk)
     recipient_email = order.get_user_current_email()
-    email_context = get_email_context(order.token)
+    email_context = get_email_base_context()
+    email_context['url'] = build_absolute_uri(
+        reverse('order:details', kwargs={'token': order.token}))
 
     # Order confirmation template requires additional information
     if template == CONFIRM_ORDER_TEMPLATE:
